utlimos cambios 8/15/24

+++ src/context/context.tsx
@@ -17,19 +17,46 @@ export const CartProvider = ({ children }: CartProviderProps) => {
   }, []);
 
   const addToCart = (product: Product) => {
-    const newCart = [...cart, product];
-    setCart(newCart);
-    localStorage.setItem('cart', JSON.stringify(newCart));
+    const existingProduct = cart.find(item => item.id === product.id);
+
+    if (existingProduct) {
+      const newCart = cart.map(item =>
+        item.id === product.id ? { ...item, quantity: item.quantity + 1 } : item
+      );
+      setCart(newCart);
+      localStorage.setItem('cart', JSON.stringify(newCart));
+    } else {
+      const newCart = [...cart, { ...product, quantity: 1 }];
+      setCart(newCart);
+      localStorage.setItem('cart', JSON.stringify(newCart));
+    }
   };
 
-  const removeFromCart = (productId: string) => {
+  const removeFromCart = (productId: number) => {
     const newCart = cart.filter(product => product.id !== productId);
     setCart(newCart);
     localStorage.setItem('cart', JSON.stringify(newCart));
   };
 
+  const updateQuantity = (productId: number, change: number) => {
+    const newCart = cart.map(item => {
+      if (item.id === productId) {
+        const newQuantity = item.quantity + change;
+        if (newQuantity > 0) {
+          return { ...item, quantity: newQuantity };
+        } else {
+          return null; // Retorna null si la cantidad es 0 o menor
+        }
+      }
+      return item;
+    }).filter(item => item !== null) as Product[];
+  
+    setCart(newCart);
+    localStorage.setItem('cart', JSON.stringify(newCart));
+  };
+  
   return (
-    <CartContext.Provider value={{ cart, addToCart, removeFromCart }}>
+    <CartContext.Provider value={{ cart, addToCart, removeFromCart, updateQuantity }}>
       {children}
     </CartContext.Provider>
   );
